refactor: reduce coupling
