+++ estudantes/yagosansz/17-reading_files/blogposts/blogposts.go
@@ -1,13 +1,8 @@
 package blogposts
 
 import (
-	"io"
 	"io/fs"
 )
-
-type Post struct {
-	Title string
-}
 
 func NewPostsFromFS(fileSystem fs.FS) ([]Post, error) {
 	dir, err := fs.ReadDir(fileSystem, ".")
@@ -16,7 +11,7 @@ func NewPostsFromFS(fileSystem fs.FS) ([]Post, error) {
 	}
 	var posts []Post
 	for _, f := range dir {
-		post, err := getPost(fileSystem, f)
+		post, err := getPost(fileSystem, f.Name())
 		if err != nil {
 			//todo: needs clarification, should we totally fail if one file fails? or just ignore?
 			return nil, err
@@ -26,19 +21,11 @@ func NewPostsFromFS(fileSystem fs.FS) ([]Post, error) {
 	return posts, nil
 }
 
-func getPost(fileSystem fs.FS, f fs.DirEntry) (Post, error) {
-	postFile, err := fileSystem.Open(f.Name())
+func getPost(fileSystem fs.FS, fileName string) (Post, error) {
+	postFile, err := fileSystem.Open(fileName)
 	if err != nil {
 		return Post{}, err
 	}
 	defer postFile.Close()
-
-	postData, err := io.ReadAll(postFile)
-	if err != nil {
-		return Post{}, err
-	}
-
-	// Here we are cutting out `Title: ` by slicing the string
-	post := Post{Title: string(postData)[7:]}
-	return post, nil
+	return newPost(postFile)
 }
